requirements.txt
annotated-types==0.7.0
anyio==4.3.0
blinker==1.8.2
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
python-dotenv==0.21.0
dnspython==2.6.1
exceptiongroup==1.2.1
Flask==3.0.3
gunicorn
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openai==1.20.0
packaging==24.0
pydantic==2.7.1
pydantic_core==2.18.2
pymongo==4.7.2
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.0
Werkzeug==3.0.3
